# docs/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "proto/iam.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/access-keys": {
      "get": {
        "summary": "访问密钥列表",
        "operationId": "IAM_ListAccessKeys",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListAccessKeysResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "userName",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "appId",
            "description": "可选，筛选特定应用的密钥",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "IAM"
        ]
      },
      "post": {
        "summary": "访问密钥管理\n创建访问密钥",
        "operationId": "IAM_CreateAccessKey",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1AccessKey"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateAccessKeyRequest"
            }
          }
        ],
        "tags": [
          "IAM"
        ]
      }
    },
...
    "/v1/applications": {
      "get": {
        "summary": "应用列表",
        "operationId": "IAM_ListApplications",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListApplicationsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "userName",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "status",
            "description": "active/inactive/suspended",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "IAM"
        ]
      },
      "post": {
        "summary": "应用管理\n创建应用",
        "operationId": "IAM_CreateApplication",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Application"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateApplicationRequest"
            }
          }
        ],
        "tags": [
          "IAM"
        ]
      }
    },
    "/v1/applications/{appId}": {
      "get": {
        "summary": "获取应用",
        "operationId": "IAM_GetApplication",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Application"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "IAM"
        ]
      },
      "delete": {
        "summary": "删除应用",
        "operationId": "IAM_DeleteApplication",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteApplicationResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "IAM"
        ]
      },
      "put": {
        "summary": "更新应用",
        "operationId": "IAM_UpdateApplication",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Application"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/IAMUpdateApplicationBody"
            }
          }
        ],
        "tags": [
          "IAM"
        ]
      }
    },
...
    "/v1/policies": {
      "get": {
        "summary": "获取策略列表",
        "operationId": "IAM_ListPolicies",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListPoliciesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "page",
            "description": "页码（从1开始）",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageSize",
            "description": "每页大小（最大100）",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "IAM"
        ]
      },
      "post": {
        "summary": "策略管理\n创建策略",
        "operationId": "IAM_CreatePolicy",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Policy"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreatePolicyRequest"
            }
          }
        ],
        "tags": [
          "IAM"
        ]
      }
    },
    "/v1/sts/assume-role": {
      "post": {
        "summary": "STS临时授权服务\n假设角色",
        "operationId": "IAM_AssumeRole",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1AssumeRoleResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1AssumeRoleRequest"
            }
          }
        ],
        "tags": [
          "IAM"
        ]
      }
    },
...
    "/v1/users": {
      "get": {
        "summary": "用户列表",
        "operationId": "IAM_ListUsers",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListUsersResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "IAM"
        ]
      },
      "post": {
        "summary": "创建用户",
        "operationId": "IAM_CreateUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1User"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateUserRequest"
            }
          }
        ],
        "tags": [
          "IAM"
        ]
      }
    },
    "/v1/users/{name}": {
      "get": {
        "summary": "获取用户",
        "operationId": "IAM_GetUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1User"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "IAM"
        ]
      },
      "delete": {
        "summary": "删除用户",
        "operationId": "IAM_DeleteUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "IAM"
        ]
      },
      "put": {
        "summary": "更新用户",
        "operationId": "IAM_UpdateUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1User"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/IAMUpdateUserBody"
            }
          }
        ],
        "tags": [
          "IAM"
        ]
      }
    },
...
  },
  "definitions": {
...
    "IAMUpdateApplicationBody": {
      "type": "object",
      "properties": {
        "appName": {
          "type": "string"
        },
        "appDescription": {
          "type": "string"
        },
        "appType": {
          "type": "string"
        },
        "appIconUrl": {
          "type": "string"
        },
        "appWebsite": {
          "type": "string"
        },
        "callbackUrls": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "allowedOrigins": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "status": {
          "type": "string"
        }
      }
    },
    "IAMUpdateUserBody": {
      "type": "object",
      "properties": {
        "displayName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    },
...
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1AccessKey": {
      "type": "object",
      "properties": {
        "accessKeyId": {
          "type": "string"
        },
        "secretAccessKey": {
          "type": "string",
          "title": "仅在创建时返回"
        },
        "status": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "appId": {
          "type": "string",
          "format": "int64"
        },
        "appName": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "v1Application": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "userName": {
          "type": "string"
        },
        "appName": {
          "type": "string"
        },
        "appDescription": {
          "type": "string"
        },
        "appType": {
          "type": "string"
        },
        "appIconUrl": {
          "type": "string"
        },
        "appWebsite": {
          "type": "string"
        },
        "callbackUrls": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "allowedOrigins": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "status": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "v1AssumeRoleRequest": {
      "type": "object",
      "properties": {
        "roleArn": {
          "type": "string",
          "title": "角色ARN"
        },
        "roleSessionName": {
          "type": "string",
          "title": "会话名称"
        },
        "durationSeconds": {
          "type": "integer",
          "format": "int32",
          "title": "有效期（秒），默认3600，最大43200"
        },
        "externalId": {
          "type": "string",
          "title": "外部ID（可选）"
        },
        "policy": {
          "type": "string",
          "title": "会话策略（可选，JSON格式）"
        },
        "tags": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "title": "会话标签（可选）"
        }
      },
      "title": "STS临时授权相关消息"
    },
    "v1AssumeRoleResponse": {
      "type": "object",
      "properties": {
        "credentials": {
          "$ref": "#/definitions/v1TemporaryCredentials",
          "title": "临时凭证"
        },
        "assumedRoleUser": {
          "$ref": "#/definitions/v1AssumedRoleUser",
          "title": "假设角色用户信息"
        },
        "packedPolicySize": {
          "type": "integer",
          "format": "int32",
          "title": "打包策略大小"
        }
      }
    },
...
    "v1CreateAccessKeyRequest": {
      "type": "object",
      "properties": {
        "userName": {
          "type": "string"
        },
        "appId": {
          "type": "string",
          "format": "int64",
          "title": "关联的应用ID"
        },
        "description": {
          "type": "string"
        }
      },
      "title": "访问密钥相关消息"
    },
    "v1CreateApplicationRequest": {
      "type": "object",
      "properties": {
        "appName": {
          "type": "string"
        },
        "appDescription": {
          "type": "string"
        },
        "appType": {
          "type": "string",
          "title": "web/mobile/desktop/api"
        },
        "appIconUrl": {
          "type": "string"
        },
        "appWebsite": {
          "type": "string"
        },
        "callbackUrls": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "allowedOrigins": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "title": "应用管理相关消息"
    },
    "v1CreatePolicyRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "policyDocument": {
          "type": "string",
          "title": "JSON字符串"
        }
      },
      "title": "策略相关消息"
    },
    "v1CreateUserRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "displayName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      },
      "title": "用户相关消息"
    },
...
    "v1DeleteApplicationResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        }
      }
    },
    "v1DeleteUserResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        }
      }
    },
...
    "v1ListAccessKeysResponse": {
      "type": "object",
      "properties": {
        "accessKeys": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1AccessKey"
          }
        }
      }
    },
    "v1ListApplicationsResponse": {
      "type": "object",
      "properties": {
        "applications": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Application"
          }
        },
        "total": {
          "type": "integer",
          "format": "int32"
        },
        "page": {
          "type": "integer",
          "format": "int32"
        },
        "pageSize": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "v1ListPoliciesResponse": {
      "type": "object",
      "properties": {
        "policies": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Policy"
          },
          "title": "策略列表"
        },
        "total": {
          "type": "integer",
          "format": "int32",
          "title": "总数"
        },
        "page": {
          "type": "integer",
          "format": "int32",
          "title": "当前页码"
        },
        "pageSize": {
          "type": "integer",
          "format": "int32",
          "title": "每页大小"
        }
      },
      "title": "策略列表响应"
    },
    "v1ListUsersResponse": {
      "type": "object",
      "properties": {
        "users": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1User"
          }
        },
        "total": {
          "type": "integer",
          "format": "int32"
        },
        "page": {
          "type": "integer",
          "format": "int32"
        },
        "size": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "v1Policy": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "policyDocument": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "v1User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "displayName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
  }
}